Fix problems on Windows (Visual Studio)
* Get the source/cpp code to compile under Visual Studio
* Fix configuration for tests for VisualStudio
* Download and compile GTest at compile time
* Fix an out-of-bounds access in Windows
* Fix setup script on Windows.

+++ source/cpp/common/source/DataStructure.cpp
@@ -51,7 +51,8 @@ class DataStructure<BaseType>::DataStructureImpl
     host_layout_factory_ = data_layout_factory;
 
     host_data_layout_ = std::unique_ptr<amsla::common::DataLayoutInterface>(
-        data_layout_factory(row_indices, column_indices, values, num_elements));
+        data_layout_factory(row_indices, column_indices, values,
+                            static_cast<uint>(num_elements)));
 
     exportable_sources_ = std::make_unique<amsla::common::DeviceSource>(
         host_data_layout_->exportDeviceSources());
